Update the magic recipe to anaylyze xsd types
