+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@ exclude: tests/fixtures
 
 repos:
   - repo: https://github.com/asottile/pyupgrade
-    rev: v3.3.1
+    rev: v3.4.0
     hooks:
       - id: pyupgrade
         args: [ --py37-plus ]
@@ -15,7 +15,7 @@ repos:
     hooks:
       - id: black
   - repo: https://github.com/PyCQA/autoflake
-    rev: v2.1.0
+    rev: v2.1.1
     hooks:
       - id: autoflake
   - repo: https://github.com/PyCQA/flake8
@@ -36,7 +36,7 @@ repos:
       - id: end-of-file-fixer
       - id: debug-statements
   - repo: https://github.com/PyCQA/docformatter
-    rev: v1.5.1
+    rev: v1.6.5
     hooks:
       - id: docformatter
         args: [ "--in-place", "--pre-summary-newline" ]
