+++ xsdata/codegen/handlers/__init__.py
@@ -1,4 +1,5 @@
 from .add_attribute_substitutions import AddAttributeSubstitutions
+from .calculate_attribute_paths import CalculateAttributePaths
 from .create_compound_fields import CreateCompoundFields
 from .designate_class_packages import DesignateClassPackages
 from .filter_classes import FilterClasses
@@ -9,6 +10,7 @@ from .process_attributes_types import ProcessAttributeTypes
 from .process_mixed_content_class import ProcessMixedContentClass
 from .rename_duplicate_attributes import RenameDuplicateAttributes
 from .rename_duplicate_classes import RenameDuplicateClasses
+from .reset_attribute_sequences import ResetAttributeSequences
 from .sanitize_attributes_default_value import SanitizeAttributesDefaultValue
 from .sanitize_enumeration_class import SanitizeEnumerationClass
 from .unnest_inner_classes import UnnestInnerClasses
@@ -18,6 +20,7 @@ from .validate_attributes_overrides import ValidateAttributesOverrides
 
 __all__ = [
     "AddAttributeSubstitutions",
+    "CalculateAttributePaths",
     "CreateCompoundFields",
     "DesignateClassPackages",
     "FilterClasses",
@@ -28,6 +31,7 @@ __all__ = [
     "ProcessMixedContentClass",
     "RenameDuplicateAttributes",
     "RenameDuplicateClasses",
+    "ResetAttributeSequences",
     "SanitizeAttributesDefaultValue",
     "SanitizeEnumerationClass",
     "UnnestInnerClasses",
